Move all activators into single file
